backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
torch==2.3.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
evidently==0.4.30
langchain==0.2.6
langgraph==0.1.14
langchain-anthropic==0.1.15
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.4
